move crcUtils, logger, utilities to a utils submodule
